Adds feature testing to ICU 63 and 64
Feature testing is needed to ensure that older generated code supports
newly added `rust_icu_*` crates.

- Ran `make static-bindgen` to refresh the bindings for all supported
  ICU versions.
- Added ICU versions 63 and 64 to the feature testing matrix.
- Generated lib_63.rs, lib_64.rs and symlinked lib.rs to it in
  rust_icu_sys.  This makes the bindgen generated code uniform across
  versions.

Fixes: 111
