+++ README.md
@@ -97,6 +97,8 @@ for those versions.
 
 ## Feature sets
 
+For an explanation of features, see the [Features section](#features) below.
+
 * 1: default
 * 2: "renaming"
 * 3: "icu_version_in_env"
@@ -109,9 +111,10 @@ this particular ICU library and `rust_icu` version combination.
 | `rust_icu` version | ICU 63.x | ICU 64.2 | ICU 65.1  | ICU 66.0.1 | ICU 67.1 |
 | ------------------ | -------- | -------- | --------- | ---------- | -------- |
 | 0.1                | ☟        | ☟        | ☟         | ☟          | ☟        |
-| 0.1.3              | 1        | 1        | 1; 2; 2+3 | 1          | 1        |
+| 0.1.3              | 1        | 1        | 1;2;2+3   | 1          | 1        |
 | 0.2                | ☟        | ☟        | ☟         | ☟          | ☟        |
-| 0.2.2              | 1        | 1        | 1; 2; 2+3 | 1          | 1        |
+| 0.2.2              | 1        | 1        | 1;2;2+3   | 1          | 1        |
+| 0.2.3              | 1;2;2+3  | 1;2;2+3  | 1;2;2+3   | 1;2;2+3    | 1;2;2+3  |
 
 > Prior to a 1.0.0 release, API versions that only differ in the patch version
 > number (0.x.**y**) only should be compatible.
@@ -137,9 +140,9 @@ confusing compilation end result.
 | Feature | Default? | Description |
 | ------- | -------- | ----------- |
 | `bindgen` | Yes | If set, cargo will run `bindgen` to generate bindings based on the installed ICU library.  The program `icu-config` must be in $PATH for this to work. In the future there may be other approaches for auto-detecting libraries, such as via `pkg-config`. |
-| `renaming` | Yes | If set, ICU bindings are generated with version numbers appended.  This is called "renaming" in ICU, and is normally needed only when linking against specific ICU version is required, for example to work around having to link different ICU versions.  See [the ICU documentation](http://userguide.icu-project.org/design) for a discussion of renaming. |
-| `icu_config` | Yes | If set, the binary icu-config will be used to configure the library.  Turn this feature off if you do not want `build.rs` to try to autodetect the build environment.  You will want to skip this feature if your build environment configures ICU in a different way. |
-| `icu_version_in_env` | No | If set, ICU bindings are made for the ICU version specified in the environment variable `RUST_ICU_MAJOR_VERSION_NUMBER`, which is made available to cargo at build time. See section below for details on how to use this feature. |
+| `renaming` | Yes | If set, ICU bindings are generated with version numbers appended.  This is called "renaming" in ICU, and is normally needed only when linking against specific ICU version is required, for example to work around having to link different ICU versions.  See [the ICU documentation](http://userguide.icu-project.org/design) for a discussion of renaming. **This feature MUST be used when `bindgen` is NOT used.** |
+| `icu_config` | Yes | If set, the binary icu-config will be used to configure the library.  Turn this feature off if you do not want `build.rs` to try to autodetect the build environment.  You will want to skip this feature if your build environment configures ICU in a different way. **This feature is only meaningful when `bindgen` feature is used; otherwise it has no effect.** |
+| `icu_version_in_env` | No | If set, ICU bindings are made for the ICU version specified in the environment variable `RUST_ICU_MAJOR_VERSION_NUMBER`, which is made available to cargo at build time. See section below for details on how to use this feature. **This feature is only meaningful when `bindgen` feature is NOT used; otherwise it has no effect.** |
 
 # Prerequisites
 
@@ -502,7 +505,13 @@ Some clarification:
 
 Requires docker.
 
-Run `make static-bindgen` periodically, to refresh the static bindgen files.
+Run `make static-bindgen` periodically, to refresh the statically generated
+bindgen files (named `lib_XX.rs`, where `XX` is an ICU version, e.g. 67) in the
+directory [`rust_icu_sys/bindgen`](./rust_icu_sys/bindgen) which are used when
+`bindgen` features are turned off.
+
+Invoking this make target will modify the local checkout with the newer versions
+of the files `lib_XX.rs`.  Make a pull request and check them in.
 
 For more information on why this is needed, see the [bindgen
 README.md](rust_icu_sys/bindgen/README.md).
