+++ rust_icu/Cargo.toml
@@ -6,7 +6,7 @@ license = "Apache-2.0"
 name = "rust_icu"
 readme = "README.md"
 repository = "https://github.com/google/rust_icu"
-version = "0.2.2"
+version = "0.2.3"
 
 description = """
 Native bindings to the ICU4C library from Unicode.
@@ -18,17 +18,17 @@ keywords = ["icu", "unicode", "i18n", "l10n"]
 anyhow = "1.0.25"
 log = "0.4.6"
 paste = "0.1.5"
-rust_icu_common = { path = "../rust_icu_common", version = "0.2.2", default-features = false }
-rust_icu_sys = { path = "../rust_icu_sys", version = "0.2.2", default-features = false }
-rust_icu_ucal = { path = "../rust_icu_ucal", version = "0.2.2", default-features = false }
-rust_icu_udat = { path = "../rust_icu_udat", version = "0.2.2", default-features = false }
-rust_icu_udata = { path = "../rust_icu_udata", version = "0.2.2", default-features = false }
-rust_icu_uenum = { path = "../rust_icu_uenum", version = "0.2.2", default-features = false }
-rust_icu_uloc = { path = "../rust_icu_uloc", version = "0.2.2", default-features = false }
-rust_icu_ucol = { path = "../rust_icu_ucol", version = "0.2.2", default-features = false }
-rust_icu_umsg = { path = "../rust_icu_umsg", version = "0.2.2", default-features = false }
-rust_icu_ustring = { path = "../rust_icu_ustring", version = "0.2.2", default-features = false }
-rust_icu_utext = { path = "../rust_icu_utext", version = "0.2.2", default-features = false }
+rust_icu_common = { path = "../rust_icu_common", version = "0.2.3", default-features = false }
+rust_icu_sys = { path = "../rust_icu_sys", version = "0.2.3", default-features = false }
+rust_icu_ucal = { path = "../rust_icu_ucal", version = "0.2.3", default-features = false }
+rust_icu_udat = { path = "../rust_icu_udat", version = "0.2.3", default-features = false }
+rust_icu_udata = { path = "../rust_icu_udata", version = "0.2.3", default-features = false }
+rust_icu_uenum = { path = "../rust_icu_uenum", version = "0.2.3", default-features = false }
+rust_icu_uloc = { path = "../rust_icu_uloc", version = "0.2.3", default-features = false }
+rust_icu_ucol = { path = "../rust_icu_ucol", version = "0.2.3", default-features = false }
+rust_icu_umsg = { path = "../rust_icu_umsg", version = "0.2.3", default-features = false }
+rust_icu_ustring = { path = "../rust_icu_ustring", version = "0.2.3", default-features = false }
+rust_icu_utext = { path = "../rust_icu_utext", version = "0.2.3", default-features = false }
 thiserror = "1.0.9"
 
 # See the feature description in ../rust_icu_sys/Cargo.toml for details.
