+++ rust_icu_sys/bindgen/run_bindgen.sh
@@ -62,6 +62,7 @@ readonly BINDGEN_ALLOWLIST_TYPES=(
         "UParseError"
         "UText"
         "UCollator"
+        "UCollation.*"
         "USet"
         "UCol.*"
 )
